+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Updated
+
+- Improvements to error reporting including for fetch failures.
+
 ## 0.6.1 - 2026-07-28
 
 ### Updated